feat: Move Mixins to separate file
